Change logic AppointmentControllerUnitTest, EntityControllerUnitTest.…

+++ src/main/java/com/example/demo/controllers/AppointmentController.java
@@ -1,22 +1,15 @@
 package com.example.demo.controllers;
 
-import com.example.demo.repositories.*;
-import com.example.demo.entities.*;
+import com.example.demo.entities.Appointment;
+import com.example.demo.repositories.AppointmentRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
 
 
 @RestController
@@ -27,12 +20,11 @@ public class AppointmentController {
     AppointmentRepository appointmentRepository;
 
     @GetMapping("/appointments")
-    public ResponseEntity<List<Appointment>> getAllAppointments(){
-        List<Appointment> appointments = new ArrayList<>();
+    public ResponseEntity<List<Appointment>> getAllAppointments() {
 
-        appointmentRepository.findAll().forEach(appointments::add);
+        List<Appointment> appointments = new ArrayList<>(appointmentRepository.findAll());
 
-        if (appointments.isEmpty()){
+        if (appointments.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
 
@@ -40,43 +32,55 @@ public class AppointmentController {
     }
 
     @GetMapping("/appointments/{id}")
-    public ResponseEntity<Appointment> getAppointmentById(@PathVariable("id") long id){
+    public ResponseEntity<Appointment> getAppointmentById(@PathVariable("id") long id) {
         Optional<Appointment> appointment = appointmentRepository.findById(id);
 
-        if (appointment.isPresent()){
-            return new ResponseEntity<>(appointment.get(),HttpStatus.OK);
-        }else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
+        return appointment.map(value -> new ResponseEntity<>(value, HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
     @PostMapping("/appointment")
-    public ResponseEntity<List<Appointment>> createAppointment(@RequestBody Appointment appointment){
-        /** TODO 
-         * Implement this function, which acts as the POST /api/appointment endpoint.
-         * Make sure to check out the whole project. Specially the Appointment.java class
-         */
-        return new ResponseEntity<>(HttpStatus.I_AM_A_TEAPOT);
+    public ResponseEntity<List<Appointment>> createAppointment(@RequestBody Appointment appointment) {
+
+        List<Appointment> appointments;
+
+        if (appointment.getStartsAt().isEqual(appointment.getFinishesAt()) ||
+                appointment.getStartsAt().isAfter(appointment.getFinishesAt())) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
+        appointments = appointmentRepository.findAll();
+        for (Appointment a : appointments) {
+            if (a.overlaps(appointment)) {
+                return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+            }
+        }
+
+        appointmentRepository.save(appointment);
+        appointments = appointmentRepository.findAll();
+
+        return new ResponseEntity<>(appointments, HttpStatus.OK);
     }
 
-
     @DeleteMapping("/appointments/{id}")
-    public ResponseEntity<HttpStatus> deleteAppointment(@PathVariable("id") long id){
+    public ResponseEntity<HttpStatus> deleteAppointment(@PathVariable("id") long id) {
 
         Optional<Appointment> appointment = appointmentRepository.findById(id);
 
-        if (!appointment.isPresent()){
+        if (!appointment.isPresent()) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
 
         appointmentRepository.deleteById(id);
-
         return new ResponseEntity<>(HttpStatus.OK);
-        
     }
 
     @DeleteMapping("/appointments")
-    public ResponseEntity<HttpStatus> deleteAllAppointments(){
+    public ResponseEntity<HttpStatus> deleteAllAppointments() {
+
+        if (appointmentRepository.count() == 0) {
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+
         appointmentRepository.deleteAll();
         return new ResponseEntity<>(HttpStatus.OK);
     }
